+++ frontend/app.vue
@@ -7,3 +7,8 @@
     <LayoutAppFooter />
   </div>
 </template>
+
+<script setup lang="ts">
+  const { init } = useDyslexicFont()
+  onMounted(init)
+</script>